chore: move subcommands where they belong
