+++ cmd/run/BUILD.bazel
@@ -6,12 +6,12 @@ go_library(
     importpath = "github.com/unkeyed/unkey/cmd/run",
     visibility = ["//visibility:public"],
     deps = [
-        "//cmd/api",
-        "//cmd/ctrl",
-        "//cmd/frontline",
-        "//cmd/krane",
-        "//cmd/sentinel",
-        "//cmd/vault",
+        "//cmd/run/api",
+        "//cmd/run/ctrl",
+        "//cmd/run/frontline",
+        "//cmd/run/krane",
+        "//cmd/run/sentinel",
+        "//cmd/run/vault",
         "//pkg/cli",
     ],
 )
