+++ cmd/run/krane/BUILD.bazel
@@ -3,7 +3,7 @@ load("@rules_go//go:def.bzl", "go_library")
 go_library(
     name = "krane",
     srcs = ["main.go"],
-    importpath = "github.com/unkeyed/unkey/cmd/krane",
+    importpath = "github.com/unkeyed/unkey/cmd/run/krane",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/cli",
