+++ main.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/unkeyed/unkey/cmd/deploy"
 	dev "github.com/unkeyed/unkey/cmd/dev"
-	"github.com/unkeyed/unkey/cmd/frontline"
 	"github.com/unkeyed/unkey/cmd/healthcheck"
 	"github.com/unkeyed/unkey/cmd/run"
 	"github.com/unkeyed/unkey/cmd/version"
@@ -29,7 +28,6 @@ func main() {
 			version.Cmd,
 			deploy.Cmd,
 			healthcheck.Cmd,
-			frontline.Cmd,
 			dev.Cmd,
 		},
 	}
